model_serving/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pillow>=7.1.2
torchserve>=0.4.0
torch-model-archiver>=0.4.0
torch-workflow-archiver>=0.1.0
numpy>=1.23.0
pyyaml>=5.3.1
requests>=2.23.0
opencv-python>=4.5.0
